+++ Editor/TransFront/Entry.cs
@@ -17,7 +17,8 @@ namespace ResoniteImportHelper.TransFront
             bool runVRCSDKPipeline,
             // ReSharper disable once InconsistentNaming
             bool runNDMF,
-            bool bakeTexture
+            bool bakeTexture,
+            bool generateIntermediateArtifact
         )
         {
             Profiler.BeginSample("PerformConversion");
@@ -33,7 +34,7 @@ namespace ResoniteImportHelper.TransFront
             
             Debug.Log("Exporting model as glTF");
             var serialized = SerializationService.ExportToAssetFolder(
-                new SerializationConfiguration(target, unmodifiableRoot, rootAlloc)
+                new SerializationConfiguration(target, unmodifiableRoot, generateIntermediateArtifact, rootAlloc)
             );
             
             Debug.Log("done");